+++ docs/project-overview.md
@@ -50,7 +50,7 @@ Next.js App Router SPA, TypeScript, Tailwind CSS. All rendering is client-side.
 |---|---|
 | `src/app/` | Next.js App Router entry points and providers |
 | `src/components/` | View components, charts, layout, feature folders, UI primitives |
-| `src/components/features/` | Feature-owned view boundaries such as overview, file upload, client versions, and user details |
+| `src/components/features/` | Feature-owned view boundaries such as overview, file upload, languages, client versions, and user details |
 | `src/components/layout/routes/` | Typed standard-route adapters/registry |
 | `src/components/charts/` | Chart.js visualizations (via react-chartjs-2) |
 | `src/domain/` | Business logic: aggregator, model config, calculators |
@@ -166,7 +166,7 @@ Phase 5 modular aggregation orchestration and Phase 8 grouped contract assembly 
 
 ### 4.3. Views
 
-`ViewRouter` (`src/components/layout/ViewRouter.tsx`) is the thin top-level coordinator for metrics-wide upload, fatal-error, and loading gates. It selects either the typed standard-route outlet or the feature-owned `UserDetailsRoute`. Each standard adapter invokes its feature selector lazily and renders the relevant presentation boundary with a narrow shared route context. `ClientVersionsView` presentation and its JetBrains/VS Code version sections are co-located under `src/components/features/client-versions/`. `UserDetailsRoute` consumes metrics, navigation, and worker contexts directly and owns selection/summary resolution, request invalidation and retry, effect-only redirects, stale-result rejection, cleanup, recoverable errors, and final `UserDetailsView` model delivery. User-details presentation, feature-only charts, day-details helpers, and route-focused tests are co-located under `src/components/features/user-details/`; pure read-model selectors remain in `src/read-models/` as the shared application projection layer. This keeps Phase 6 feature organization moving while preserving Phase 7 routing and Phase 10 dependency enforcement.
+`ViewRouter` (`src/components/layout/ViewRouter.tsx`) is the thin top-level coordinator for metrics-wide upload, fatal-error, and loading gates. It selects either the typed standard-route outlet or the feature-owned `UserDetailsRoute`. Each standard adapter invokes its feature selector lazily and renders the relevant presentation boundary with a narrow shared route context. `LanguagesView` presentation, feature-only daily chart, visible sections, and focused tests are co-located under `src/components/features/languages/`. `ClientVersionsView` presentation and its JetBrains/VS Code version sections are co-located under `src/components/features/client-versions/`. `UserDetailsRoute` consumes metrics, navigation, and worker contexts directly and owns selection/summary resolution, request invalidation and retry, effect-only redirects, stale-result rejection, cleanup, recoverable errors, and final `UserDetailsView` model delivery. User-details presentation, feature-only charts, day-details helpers, and route-focused tests are co-located under `src/components/features/user-details/`; pure read-model selectors remain in `src/read-models/` as the shared application projection layer. This keeps Phase 6 feature organization moving while preserving Phase 7 routing and Phase 10 dependency enforcement.
 
 Charts use **Chart.js** via **react-chartjs-2**, wrapped in a `ChartContainer` component for consistent styling.
