+++ src/components/features/languages/LanguagesView.tsx
@@ -1,26 +1,18 @@
 'use client';
 
-import { LanguageStats } from '../../../domain/calculators/metricCalculators';
-import { useMemo, useState } from 'react';
-import MetricsTable, { SortState as TableSortState, TableColumn } from '../../ui/MetricsTable';
-import { DashboardStatsCardGroup, ViewPanel } from '../../ui';
-import LanguageDailyChart from './charts/LanguageDailyChart';
-import { translateFeature } from '../../../domain/featureTranslations';
-import { sortBySelector, rankBySelector } from '../../../utils/sorting';
-import { LANGUAGES_SECTIONS } from '../../layout/contextSections';
+import { ViewPanel } from '../../ui';
 import type { LanguagesReadModel } from '../../../read-models/languages';
+import CompleteLanguagesBreakdownSection from './sections/CompleteLanguagesBreakdownSection';
+import LanguageDailyChartsSection from './sections/LanguageDailyChartsSection';
+import LanguageFeatureImpactSection from './sections/LanguageFeatureImpactSection';
+import LanguageSummarySection from './sections/LanguageSummarySection';
+import NetProductivityImpactSection from './sections/NetProductivityImpactSection';
+import TopLanguageListsSection from './sections/TopLanguageListsSection';
+import { LANGUAGES_SECTIONS } from './languagesSections';
 
 interface LanguagesViewProps {
   model: LanguagesReadModel;
 }
-
-type SortField = 'language' | 'totalGenerations' | 'totalAcceptances' | 'totalEngagements' | 'uniqueUsers' | 'locAdded' | 'locDeleted' | 'locSuggestedToAdd' | 'locSuggestedToDelete';
-
-const formatAcceptanceRate = (lang: LanguageStats) => {
-  return lang.totalGenerations > 0
-    ? ((lang.totalAcceptances / lang.totalGenerations) * 100).toFixed(1)
-    : '0.0';
-};
 
 export default function LanguagesView({ model }: LanguagesViewProps) {
   const {
@@ -29,343 +21,6 @@ export default function LanguagesView({ model }: LanguagesViewProps) {
     dailyLanguageGenerationsData,
     dailyLanguageLocData,
   } = model;
-  const [tableSortState, setTableSortState] = useState<TableSortState>({
-    field: 'totalEngagements',
-    direction: 'desc',
-  });
-
-  const sortSelectors = useMemo<Record<SortField, (lang: LanguageStats) => string | number>>(
-    () => ({
-      language: (lang) => lang.language.toLowerCase(),
-      totalGenerations: (lang) => lang.totalGenerations,
-      totalAcceptances: (lang) => lang.totalAcceptances,
-      totalEngagements: (lang) => lang.totalEngagements,
-      uniqueUsers: (lang) => lang.uniqueUsers,
-      locAdded: (lang) => lang.locAdded,
-      locDeleted: (lang) => lang.locDeleted,
-      locSuggestedToAdd: (lang) => lang.locSuggestedToAdd,
-      locSuggestedToDelete: (lang) => lang.locSuggestedToDelete,
-    }),
-    [],
-  );
-
-  const sortedLanguages = useMemo(() => {
-    const field = (tableSortState.field as SortField) || 'totalEngagements';
-    const selector = sortSelectors[field];
-    return sortBySelector(languages, selector, tableSortState.direction);
-  }, [languages, sortSelectors, tableSortState]);
-
-  const handleTableSortChange = (next: TableSortState) => {
-    const field = (next.field as SortField) || 'totalEngagements';
-    setTableSortState({ field, direction: next.direction });
-  };
-
-  const totalLanguages = languages.length;
-  const totalUsers = Math.max(...languages.map(lang => lang.uniqueUsers), 0);
-  const totalLocAdded = languages.reduce((sum, lang) => sum + lang.locAdded, 0);
-  const totalLocDeleted = languages.reduce((sum, lang) => sum + lang.locDeleted, 0);
-
-  const totalNetLocImpact = totalLocAdded - totalLocDeleted;
-
-  const netLocImpactAccent: 'green' | 'rose' = totalNetLocImpact >= 0 ? 'green' : 'rose';
-
-  const summaryCards = [
-    {
-      value: totalLanguages,
-      label: 'Total Languages',
-      accent: 'blue' as const,
-    },
-    {
-      value: totalUsers,
-      label: 'Max Users/Lang',
-      accent: 'teal' as const,
-    },
-    {
-      value: totalLocAdded,
-      label: 'LOC Added',
-      accent: 'orange' as const,
-    },
-    {
-      value: totalLocDeleted,
-      label: 'LOC Deleted',
-      accent: 'rose' as const,
-    },
-    {
-      value: totalNetLocImpact,
-      label: 'Net LOC Impact',
-      accent: netLocImpactAccent,
-    },
-  ];
-
-  const languagesByGenerations = useMemo(
-    () => sortBySelector(languages, lang => lang.totalGenerations, 'desc'),
-    [languages],
-  );
-
-  const languagesByUsers = useMemo(
-    () => sortBySelector(languages, lang => lang.uniqueUsers, 'desc'),
-    [languages],
-  );
-
-  const languagesByNetLocImpact = useMemo(
-    () => sortBySelector(languages, lang => lang.locAdded - lang.locDeleted, 'desc'),
-    [languages],
-  );
-
-  const generationRankMap = useMemo(
-    () => rankBySelector(languages, lang => lang.language, lang => lang.totalGenerations),
-    [languages],
-  );
-
-  const userRankMap = useMemo(
-    () => rankBySelector(languages, lang => lang.language, lang => lang.uniqueUsers),
-    [languages],
-  );
-
-  const netImpactRankMap = useMemo(
-    () => rankBySelector(languages, lang => lang.language, lang => lang.locAdded - lang.locDeleted),
-    [languages],
-  );
-
-  const narrowHeaderClassName = 'px-4 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider';
-  const narrowHeaderRightClassName = 'px-4 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider';
-  const narrowCellClassName = 'px-4 py-4 whitespace-nowrap text-sm text-gray-900';
-  const narrowCellRightClassName = 'px-4 py-4 whitespace-nowrap text-sm text-gray-900 text-right';
-  const wideCellRightClassName = 'px-6 py-4 whitespace-nowrap text-sm text-gray-900 text-right';
-  const wideHeaderRightClassName = 'px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider';
-
-  const languagesByGenerationsColumns: TableColumn<LanguageStats>[] = [
-    {
-      id: 'language',
-      header: 'Language',
-      headerClassName: narrowHeaderClassName,
-      className: narrowCellClassName,
-      renderCell: (lang) => {
-        const rank = generationRankMap.get(lang.language) ?? '-';
-        return (
-          <div className="flex items-center">
-            <div className="flex-shrink-0 h-8 w-8 mr-3">
-              <div className="h-8 w-8 rounded bg-gradient-to-r from-blue-500 to-purple-600 flex items-center justify-center text-white text-xs font-bold">
-                {rank}
-              </div>
-            </div>
-            <div className="text-sm font-medium text-gray-900">{lang.language}</div>
-          </div>
-        );
-      },
-    },
-    {
-      id: 'generations',
-      header: 'Generations',
-      headerClassName: narrowHeaderRightClassName,
-      className: narrowCellRightClassName,
-      renderCell: (lang) => (
-        <div>{lang.totalGenerations.toLocaleString()}</div>
-      ),
-    },
-    {
-      id: 'acceptanceRate',
-      header: 'Acceptance Rate',
-      headerClassName: narrowHeaderRightClassName,
-      className: narrowCellRightClassName,
-      renderCell: (lang) => (
-        <div>{formatAcceptanceRate(lang)}%</div>
-      ),
-    },
-  ];
-
-  const languagesByUsersColumns: TableColumn<LanguageStats>[] = [
-    {
-      id: 'language',
-      header: 'Language',
-      headerClassName: narrowHeaderClassName,
-      className: narrowCellClassName,
-      renderCell: (lang) => {
-        const rank = userRankMap.get(lang.language) ?? '-';
-        return (
-          <div className="flex items-center">
-            <div className="flex-shrink-0 h-8 w-8 mr-3">
-              <div className="h-8 w-8 rounded bg-gradient-to-r from-green-500 to-teal-600 flex items-center justify-center text-white text-xs font-bold">
-                {rank}
-              </div>
-            </div>
-            <div className="text-sm font-medium text-gray-900">{lang.language}</div>
-          </div>
-        );
-      },
-    },
-    {
-      id: 'users',
-      header: 'Users',
-      headerClassName: narrowHeaderRightClassName,
-      className: narrowCellRightClassName,
-      renderCell: (lang) => (
-        <div>{lang.uniqueUsers.toLocaleString()}</div>
-      ),
-    },
-    {
-      id: 'avgEngagements',
-      header: 'Avg Engagements',
-      headerClassName: narrowHeaderRightClassName,
-      className: narrowCellRightClassName,
-      renderCell: (lang) => {
-        const avgEngagements = lang.uniqueUsers > 0
-          ? (lang.totalEngagements / lang.uniqueUsers).toFixed(1)
-          : '0.0';
-        return <div>{avgEngagements}</div>;
-      },
-    },
-  ];
-
-  const languagesByNetImpactColumns: TableColumn<LanguageStats>[] = [
-    {
-      id: 'rank',
-      header: 'Rank',
-      headerClassName: narrowHeaderClassName,
-      className: narrowCellClassName,
-      renderCell: (lang) => {
-        const rank = netImpactRankMap.get(lang.language);
-        return <div className="text-sm font-medium text-gray-900">{rank ?? '–'}</div>;
-      },
-    },
-    {
-      id: 'language',
-      header: 'Language',
-      headerClassName: narrowHeaderClassName,
-      className: narrowCellClassName,
-      renderCell: (lang) => (
-        <div className="text-sm font-medium text-gray-900">{lang.language}</div>
-      ),
-    },
-    {
-      id: 'netImpact',
-      header: 'Net LOC Impact',
-      headerClassName: narrowHeaderRightClassName,
-      className: narrowCellRightClassName,
-      renderCell: (lang) => {
-        const netLocImpact = lang.locAdded - lang.locDeleted;
-        const impactColor = netLocImpact > 0 ? 'text-green-600' : netLocImpact < 0 ? 'text-rose-600' : 'text-gray-500';
-        return (
-          <div className={`font-medium ${impactColor}`}>
-            {netLocImpact.toLocaleString()}
-          </div>
-        );
-      },
-    },
-    {
-      id: 'totalGenerations',
-      header: 'Total Generations',
-      headerClassName: narrowHeaderRightClassName,
-      className: narrowCellRightClassName,
-      renderCell: (lang) => (
-        <div>{lang.totalGenerations.toLocaleString()}</div>
-      ),
-    },
-    {
-      id: 'acceptanceRate',
-      header: 'Acceptance Rate',
-      headerClassName: narrowHeaderRightClassName,
-      className: narrowCellRightClassName,
-      renderCell: (lang) => (
-        <div>{formatAcceptanceRate(lang)}%</div>
-      ),
-    },
-  ];
-
-  const tableRowClassName = () => 'hover:bg-gray-50';
-
-  const completeLanguagesColumns: TableColumn<LanguageStats>[] = [
-    {
-      id: 'language',
-      header: 'Language',
-      sortable: true,
-      className: 'px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-900',
-      renderCell: (lang) => (
-        <div className="text-sm font-medium text-gray-900">{lang.language}</div>
-      ),
-    },
-    {
-      id: 'totalEngagements',
-      header: 'Total Engagements',
-      sortable: true,
-      accessor: 'totalEngagements',
-      headerClassName: wideHeaderRightClassName,
-      className: wideCellRightClassName,
-    },
-    {
-      id: 'totalGenerations',
-      header: 'Generations',
-      sortable: true,
-      accessor: 'totalGenerations',
-      headerClassName: wideHeaderRightClassName,
-      className: wideCellRightClassName,
-    },
-    {
-      id: 'totalAcceptances',
-      header: 'Acceptances',
-      sortable: true,
-      accessor: 'totalAcceptances',
-      headerClassName: wideHeaderRightClassName,
-      className: wideCellRightClassName,
-    },
-    {
-      id: 'uniqueUsers',
-      header: 'Unique Users',
-      sortable: true,
-      accessor: 'uniqueUsers',
-      headerClassName: wideHeaderRightClassName,
-      className: wideCellRightClassName,
-    },
-    {
-      id: 'locAdded',
-      header: 'LOC Added',
-      sortable: true,
-      accessor: 'locAdded',
-      headerClassName: wideHeaderRightClassName,
-      className: wideCellRightClassName,
-    },
-    {
-      id: 'locDeleted',
-      header: 'LOC Deleted',
-      sortable: true,
-      accessor: 'locDeleted',
-      headerClassName: wideHeaderRightClassName,
-      className: wideCellRightClassName,
-    },
-    {
-      id: 'locSuggestedToAdd',
-      header: 'Suggested Add',
-      sortable: true,
-      accessor: 'locSuggestedToAdd',
-      headerClassName: wideHeaderRightClassName,
-      className: wideCellRightClassName,
-    },
-    {
-      id: 'netLocImpact',
-      header: 'Net LOC Impact',
-      headerClassName: wideHeaderRightClassName,
-      className: wideCellRightClassName,
-      renderCell: (lang) => {
-        const netLocImpact = lang.locAdded - lang.locDeleted;
-        const impactColor = netLocImpact > 0 ? 'text-green-600' : netLocImpact < 0 ? 'text-rose-600' : 'text-gray-500';
-        return (
-          <div className={`text-sm font-medium ${impactColor}`}>
-            {netLocImpact.toLocaleString()}
-          </div>
-        );
-      },
-    },
-    {
-      id: 'acceptanceRate',
-      header: 'Acceptance Rate',
-      headerClassName: wideHeaderRightClassName,
-      className: wideCellRightClassName,
-      renderCell: (lang) => (
-        <div className="text-sm text-gray-900">{formatAcceptanceRate(lang)}%</div>
-      ),
-    },
-  ];
-  const maxItemsToShow = 10;
   const [
     summarySection,
     dailyChartsSection,
@@ -383,151 +38,19 @@ export default function LanguagesView({ model }: LanguagesViewProps) {
       }}
       contentClassName="space-y-6"
     >
-      {/* Summary Stats */}
-      <div id={summarySection.id} className="scroll-mt-28">
-        <DashboardStatsCardGroup
-          className="mb-6"
-          columns={{ base: 2, md: 5 }}
-          gapClassName="gap-4"
-          items={summaryCards}
-        />
-      </div>
-
-      {/* Daily Language Charts */}
-      <div id={dailyChartsSection.id} className="space-y-6 mt-6 pt-6 border-t border-gray-200 scroll-mt-28">
-        <LanguageDailyChart chartData={dailyLanguageGenerationsData} variant="generations" />
-        <LanguageDailyChart chartData={dailyLanguageLocData} variant="loc" />
-      </div>
-
-      {/* LOC Impact by Language and Feature */}
-      {languageFeatureImpactData.rows.length > 0 && (
-        <div className="mt-6 pt-6 border-t border-gray-200">
-          <h3 className="text-lg font-semibold text-gray-900 mb-2">LOC Impact by Language and Feature</h3>
-          <p className="text-sm text-gray-500 mb-4">
-            Lines of code impacted (added + deleted) for top 10 languages, broken down by Copilot feature.
-          </p>
-          <div className="overflow-x-auto border border-gray-200">
-            <table className="w-full divide-y divide-gray-200">
-              <thead className="bg-gray-50">
-                <tr>
-                  <th className="px-4 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Language
-                  </th>
-                  <th className="px-4 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider">
-                    Total
-                  </th>
-                  {languageFeatureImpactData.features.map((feature) => (
-                    <th
-                      key={feature}
-                      className="px-4 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider"
-                    >
-                      {translateFeature(feature)}
-                    </th>
-                  ))}
-                </tr>
-              </thead>
-              <tbody className="bg-white divide-y divide-gray-200">
-                {languageFeatureImpactData.rows.map((row, index) => (
-                  <tr key={row.language} className={index % 2 === 0 ? 'bg-white' : 'bg-gray-50'}>
-                    <td className="px-4 py-3 whitespace-nowrap text-sm font-medium text-gray-900">
-                      {row.language}
-                    </td>
-                    <td className="px-4 py-3 whitespace-nowrap text-sm text-gray-900 text-right font-medium">
-                      {row.total.toLocaleString()}
-                    </td>
-                    {languageFeatureImpactData.features.map((feature) => (
-                      <td
-                        key={feature}
-                        className="px-4 py-3 whitespace-nowrap text-sm text-gray-900 text-right"
-                      >
-                        {row.features[feature].toLocaleString()}
-                      </td>
-                    ))}
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-        </div>
-      )}
-
-      {/* Two Column Layout for Tables */}
-      <div id={topListsSection.id} className="grid grid-cols-1 xl:grid-cols-2 gap-6 mt-6 pt-6 border-t border-gray-200 scroll-mt-28">
-        {/* Languages by Number of Generations */}
-        <div>
-          <h3 className="text-lg font-semibold text-gray-900 mb-4">Languages by Code Generations</h3>
-          <MetricsTable
-            data={languagesByGenerations}
-            columns={languagesByGenerationsColumns}
-            rowClassName={tableRowClassName}
-            tableClassName="w-full divide-y divide-gray-200"
-            tableContainerClassName="overflow-x-auto border border-gray-200"
-            theadClassName="bg-gray-50"
-            initialCount={maxItemsToShow}
-            buttonCollapsedLabel={(total) => `Show All ${total} Languages`}
-            buttonExpandedLabel="Show Less"
-          />
-        </div>
-
-        {/* Languages by Number of Users */}
-        <div>
-          <h3 className="text-lg font-semibold text-gray-900 mb-4">Languages by Number of Users</h3>
-          <MetricsTable
-            data={languagesByUsers}
-            columns={languagesByUsersColumns}
-            rowClassName={tableRowClassName}
-            tableClassName="w-full divide-y divide-gray-200"
-            tableContainerClassName="overflow-x-auto border border-gray-200"
-            theadClassName="bg-gray-50"
-            initialCount={maxItemsToShow}
-            buttonCollapsedLabel={(total) => `Show All ${total} Languages`}
-            buttonExpandedLabel="Show Less"
-          />
-        </div>
-      </div>
-
-      {/* Net Productivity Impact by Language */}
-      <div id={netImpactSection.id} className="mt-6 pt-6 border-t border-gray-200 scroll-mt-28">
-        <h3 className="text-lg font-semibold text-gray-900 mb-2">Net Productivity Impact by Language</h3>
-        <p className="text-sm text-gray-500 mb-4">
-          Net LOC impact estimates how much accepted code Copilot is changing per language, combining lines of code added and deleted.
-        </p>
-        <MetricsTable
-          data={languagesByNetLocImpact}
-          columns={languagesByNetImpactColumns}
-          rowClassName={tableRowClassName}
-          tableClassName="w-full divide-y divide-gray-200"
-          tableContainerClassName="overflow-x-auto border border-gray-200"
-          theadClassName="bg-gray-50"
-          initialCount={maxItemsToShow}
-          buttonCollapsedLabel={(total) => `Show All ${total} Languages`}
-          buttonExpandedLabel="Show Less"
-        />
-      </div>
-
-      {/* Full Languages Table */}
-      <div id={completeBreakdownSection.id} className="mt-6 pt-6 border-t border-gray-200 scroll-mt-28">
-        <h3 className="text-lg font-semibold text-gray-900 mb-4">Complete Languages Breakdown</h3>
-        <MetricsTable
-          data={sortedLanguages}
-          columns={completeLanguagesColumns}
-          sortState={tableSortState}
-          onSortChange={handleTableSortChange}
-          rowClassName={tableRowClassName}
-          tableClassName="w-full divide-y divide-gray-200"
-          tableContainerClassName="overflow-x-auto border border-gray-200"
-          theadClassName="bg-gray-50"
-          initialCount={maxItemsToShow}
-          buttonCollapsedLabel={(total) => `Show All ${total} Languages`}
-          buttonExpandedLabel="Show Less"
-        />
-
-        {languages.length === 0 && (
-          <div className="text-center py-8">
-            <p className="text-gray-500">No language data available</p>
-          </div>
-        )}
-      </div>
+      <LanguageSummarySection sectionId={summarySection.id} languages={languages} />
+      <LanguageDailyChartsSection
+        sectionId={dailyChartsSection.id}
+        dailyLanguageGenerationsData={dailyLanguageGenerationsData}
+        dailyLanguageLocData={dailyLanguageLocData}
+      />
+      <LanguageFeatureImpactSection languageFeatureImpactData={languageFeatureImpactData} />
+      <TopLanguageListsSection sectionId={topListsSection.id} languages={languages} />
+      <NetProductivityImpactSection sectionId={netImpactSection.id} languages={languages} />
+      <CompleteLanguagesBreakdownSection
+        sectionId={completeBreakdownSection.id}
+        languages={languages}
+      />
     </ViewPanel>
   );
 }
