+++ src/components/features/languages/index.ts
@@ -1,1 +1,2 @@
 export { default as LanguagesView } from './LanguagesView';
+export { LANGUAGES_SECTIONS } from './languagesSections';
